Organize tests in the same manner as views
